requirements.txt
biopython==1.72
numpy==1.15.3
pandas==0.23.4
patsy==0.5.1
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.7
scipy==1.1.0
six==1.11.0
statsmodels==0.9.0
